+++ packages/apollo-server-core/src/ApolloServer.ts
@@ -260,6 +260,16 @@ export class ApolloServerBase {
       // Don't add the extension here (we want to add it later in generateSchemaDerivedData).
     }
 
+    if (gateway && this.engineReportingAgent?.schemaReport) {
+      throw new Error(
+        [
+          "Schema reporting is not yet compatible with the gateway. If you're",
+          "interested in using schema reporting with the gateway, please",
+          "contact Apollo support.",
+        ].join(' '),
+      );
+    }
+
     this.playgroundOptions = createPlaygroundOptions(playground);
 
     // TODO: This is a bit nasty because the subscription server needs this.schema synchronously, for reasons of backwards compatibility.
@@ -479,9 +489,13 @@ export class ApolloServerBase {
   // survive transformations like monkey-patching a boolean field onto the
   // schema.
   //
-  // The only thing this is used for is determining whether traces should be
-  // added to responses if requested with an HTTP header; if there's a false
-  // positive, that feature can be disabled by specifying `engine: false`.
+  // This is used for two things:
+  // 1) determining whether traces should be added to responses if requested
+  //    with an HTTP header; if there's a false positive, that feature can be
+  //    disabled by specifying `engine: false`.
+  // 2) determining whether schema-reporting should be allowed; federated
+  //    services shouldn't be reporting schemas, and we accordingly throw if
+  //    it's attempted.
   private schemaIsFederated(schema: GraphQLSchema): boolean {
     const serviceType = schema.getType('_Service');
     if (!(serviceType && isObjectType(serviceType))) {
@@ -554,6 +568,16 @@ export class ApolloServerBase {
             'to report metrics to Apollo Graph Manager. You should only configure your Apollo gateway ' +
             'to report metrics to Apollo Graph Manager.',
         );
+
+        if (this.engineReportingAgent.schemaReport) {
+          throw Error(
+            [
+              "Schema reporting is not yet compatible with federated services.",
+              "If you're interested in using schema reporting with federated",
+              "services, please contact Apollo support.",
+            ].join(' '),
+          );
+        }
       }
       pluginsToInit.push(this.engineReportingAgent!.newPlugin());
     } else if (engine !== false && federatedSchema) {
@@ -633,7 +657,6 @@ export class ApolloServerBase {
         any
       >,
       parseOptions: this.parseOptions,
-      reporting: !!this.engineReportingAgent,
       ...this.requestOptions,
     };
   }