+++ packages/apollo-server-core/src/runHttpQuery.ts
@@ -185,8 +185,6 @@ export async function runHttpQuery(
     debug: options.debug,
 
     plugins: options.plugins || [],
-
-    reporting: options.reporting,
   };
 
   return processHTTPRequest(config, request);
@@ -264,9 +262,7 @@ export async function processHTTPRequest<TContext>(
       context,
       cache: options.cache,
       debug: options.debug,
-      metrics: {
-        captureTraces: !!options.reporting,
-      },
+      metrics: {},
     };
   }