+++ packages/apollo-gateway/CHANGELOG.md
@@ -4,7 +4,15 @@
 
 > The changes noted within this `vNEXT` section have not been released yet.  New PRs and commits which introduce changes should include an entry in this `vNEXT` section as part of their development.  When a release is being prepared, a new header will be (manually) created below and the appropriate changes within that release will be moved into the new section.
 
-- _Nothing yet! Stay tuned._
+- The default branch of the repository has been changed to `main`.  As this changed a number of references in the repository's `package.json` and `README.md` files (e.g., for badges, links, etc.), this necessitates a release to publish those changes to npm. [PR #4302](https://github.com/apollographql/apollo-server/pull/4302)
+
+## 0.16.9
+
+- Only changes in the similarly versioned `@apollo/federation` package.
+
+## 0.16.7
+
+- Bumped the version of `apollo-server-core`, but no other changes!
 
 ## 0.16.6
 
@@ -84,7 +92,7 @@
 
 ## v0.12.1
 
-- Update to include [fixes from `@apollo/federation`](https://github.com/apollographql/apollo-server/blob/master/packages/apollo-federation/CHANGELOG.md).
+- Update to include [fixes from `@apollo/federation`](https://github.com/apollographql/apollo-server/blob/main/packages/apollo-federation/CHANGELOG.md).
 
 ## v0.12.0